Go On Top Button
i tried to add a  button that will bring the user on the top on the page.

as i am a beginner on github, this is my first branch or fork. Even first time i have commit changes on github. please give me a feedback about it so that i can learn if i have done any mistake.